+++ lib/widgets/walletconnect_modal_connect.dart
@@ -199,10 +199,7 @@ class _WalletConnectModalConnectState extends State<WalletConnectModalConnect> {
   }
 
   void _reconnect() {
-    widget.service.web3App!.core.relayClient.connect(
-        // relayUrl: WalletConnectConstants.DEFAULT_RELAY_URL,
-        // relayUrl: 'https://test.walletconnect.org',
-        );
+    widget.service.reconnectRelay();
 
     setState(() {
       _state = WalletConnectModalConnectButtonState.reconnecting;